refactor : remove old codes
