--- src/main/java/nextstep/subway/auth/ui/interceptor/authentication/NewSessionAuthenticationInterceptor.java
@@ -1,27 +1,0 @@
-package nextstep.subway.auth.ui.interceptor.authentication;
-
-import nextstep.subway.auth.domain.Authentication;
-import nextstep.subway.auth.infrastructure.SecurityContext;
-import nextstep.subway.member.application.CustomUserDetailsService;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import java.io.IOException;
-
-import static nextstep.subway.auth.infrastructure.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
-
-public class NewSessionAuthenticationInterceptor extends AbstractAuthenticationInterceptor  {
-
-    public NewSessionAuthenticationInterceptor(AuthenticationConverter converter, CustomUserDetailsService userDetailsService) {
-        super(converter, userDetailsService);
-    }
-
-    @Override
-    protected void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        HttpSession session = request.getSession();
-        session.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
-    }
-
-}
