--- src/main/java/nextstep/subway/auth/ui/interceptor/authentication/NewTokenAuthenticationInterceptor.java
@@ -1,45 +1,0 @@
-package nextstep.subway.auth.ui.interceptor.authentication;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.subway.auth.domain.Authentication;
-import nextstep.subway.auth.dto.TokenResponse;
-import nextstep.subway.auth.infrastructure.JwtTokenProvider;
-import nextstep.subway.member.application.CustomUserDetailsService;
-import nextstep.subway.member.domain.LoginMember;
-import org.springframework.http.MediaType;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class NewTokenAuthenticationInterceptor extends AbstractAuthenticationInterceptor {
-    private final JwtTokenProvider jwtTokenProvider;
-    private final ObjectMapper objectMapper;
-
-    public NewTokenAuthenticationInterceptor(AuthenticationConverter converter, CustomUserDetailsService customUserDetailsService, JwtTokenProvider jwtTokenProvider, ObjectMapper objectMapper) {
-        super(converter, customUserDetailsService);
-        this.jwtTokenProvider = jwtTokenProvider;
-        this.objectMapper = objectMapper;
-    }
-
-    @Override
-    protected void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        String token = getToken(authentication);
-        TokenResponse tokenResponse = new TokenResponse(token);
-
-        setSuccessfulResponse(response, tokenResponse);
-    }
-
-    private void setSuccessfulResponse(HttpServletResponse response, TokenResponse tokenResponse) throws IOException {
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.getWriter().write(objectMapper.writeValueAsString(tokenResponse));
-        response.flushBuffer();
-    }
-
-    private String getToken(Authentication authentication) {
-        LoginMember loginMember = (LoginMember) authentication.getPrincipal();
-        return jwtTokenProvider.createToken(loginMember.getEmail());
-    }
-
-}
