+++ src/test/java/nextstep/subway/auth/ui/interceptor/authentication/SessionAuthenticationInterceptorTest.java
@@ -44,7 +44,7 @@ class SessionAuthenticationInterceptorTest {
         //given
         request = new MockHttpServletRequest();
         response = new MockHttpServletResponse();
-        interceptor = new NewSessionAuthenticationInterceptor(converter, userDetailsService);
+        interceptor = new SessionAuthenticationInterceptor(converter, userDetailsService);
     }
 
     @DisplayName("Basic방식으로 인증 실패시 에러를 던진다.")
